refactor(ts): extract raw Embind type surface to raw-types.ts
index.ts had ~350 lines of pure type declarations (OcctWasmModule, OcctRawKernel,
the Raw*/Embind* structs) ahead of any logic. Move them to raw-types.ts and
import back the ones index.ts uses; re-export OcctWasmModule/OcctRawKernel so the
public API (getRawModule/getRawKernel return types) is unchanged.

Pure type move — no runtime change. index.ts drops 2483 -> 2128 lines.
